inital react code for chat component and messages component
